Update to allow flash to load when not visible
Refresh a page containing this directive and change tabs in your browser before it is done loading. The previous timeout event would fire and the evt.ref.PercentLoaded was undefined. The timeout would not continue and the swfLoad event would never load. This change just removes the timeout and sets the interval so it will continue to check.